requirements.txt
streamlit==1.31.0
pymongo==4.6.1
python-dotenv==1.0.0
openai==1.9.0
pdfminer.six==20221105
